Use generated source not jarjar
This is in preparation to adding UnsupportedAppUsage annotations to the
source.

Note: Although the classes in the android/ directory are in the correct
package and so do not need to be moved to a different package they do
reference classes that are in the wrong package and so they need
those references transformed.

Tested using:
  m -j32 droid
  adb reboot bootloader
  fastboot flashall -w
  m -j32 cts
  cts-tradefed run cts -m CtsTelephonyTestCases -t android.telephony.cts.PhoneNumberUtilsTest

Bug: 117818301
Test: see above
Change-Id: I523e98eabc7eddff593afeddd399c11061664a59