Annotate repackaged API with @UnsupportedAppUsage
Adds @UnsupportedAppUsage annotations to the parts of libphonenumber
that are in the hidden API. Allows the libphonenumber entries to be
removed from the frameworks/base/config/hiddenapi-light-greylist.txt.

The build rule had to be changed to depend (by default) on the
core.platform.api.stubs as UnsupportedAppUsage annotation is not
present in the core_current API.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Bug: 117818301
Test: see above
Change-Id: Ic8c6a424a083b1a6c567126b93f49ce10f4e5f93