Switch implicit hidden api methods to use UnsupportedAppUsage
Implicit methods are not present in the source and so have no place on
which to add an annotation. Instead UnsupportedAppUsage annotations
are added to the class and use the implicitMember property to specify
the signature of the implicit members.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Bug: 117818301
Bug: 119861512
Test: as above
Change-Id: I904c34985baab6b80bbc3c1cfd576a4dc782a023